requirements.txt
appnope==0.1.2
asgiref==3.4.1
backcall==0.2.0
certifi==2021.10.8
charset-normalizer==2.0.9
decorator==5.1.0
Django==4.0
django-json-widget==1.1.1
djangorestframework==3.12.4
dnspython==2.1.0
filelock==3.4.0
future==0.18.2
idna==3.3
ipdb==0.13.9
ipython==7.30.1
jedi==0.18.1
matplotlib-inline==0.1.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.23
ptyprocess==0.7.0
py3-validate-email==1.0.5
Pygments==2.10.0
pytz==2021.3
requests==2.26.0
sqlparse==0.4.2
toml==0.10.2
traitlets==5.1.1
urllib3==1.26.7
wcwidth==0.2.5
